Phase-2/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend API for Task Management"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.109.0",
    "sqlmodel>=0.0.14",
    "PyJWT>=2.8.0",
    "cryptography>=41.0.0",
    "httpx>=0.26.0",
    "pydantic-settings>=2.2.0",
    "uvicorn[standard]>=0.27.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "asyncpg>=0.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.12.0",
    "httpx>=0.26.0",
    "ruff>=0.1.0"
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
